move totp related functions
